Move projectList
